Backend Server/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.txt
#
asgiref==3.5.0
    # via
    #   -r requirements.txt
    #   django
defusedxml==0.7.1
    # via
    #   -r requirements.txt
    #   odfpy
diff-match-patch==20200713
    # via
    #   -r requirements.txt
    #   django-import-export
django==4.0.3
    # via
    #   -r requirements.txt
    #   django-import-export
    #   djangorestframework
django-import-export==2.8.0
    # via -r requirements.txt
djangorestframework==3.13.1
    # via -r requirements.txt
et-xmlfile==1.1.0
    # via
    #   -r requirements.txt
    #   openpyxl
gunicorn==20.1.0
    # via -r requirements.txt
markuppy==1.14
    # via
    #   -r requirements.txt
    #   tablib
odfpy==1.4.1
    # via
    #   -r requirements.txt
    #   tablib
openpyxl==3.0.9
    # via
    #   -r requirements.txt
    #   tablib
psycopg2==2.9.3
    # via -r requirements.txt
pytz==2022.1
    # via
    #   -r requirements.txt
    #   djangorestframework
pyyaml==6.0
    # via
    #   -r requirements.txt
    #   tablib
sqlparse==0.4.2
    # via
    #   -r requirements.txt
    #   django
tablib[html,ods,xls,xlsx,yaml]==3.2.0
    # via
    #   -r requirements.txt
    #   django-import-export
tzdata==2022.1
    # via
    #   -r requirements.txt
    #   django
whitenoise==6.0.0
    # via -r requirements.txt
xlrd==2.0.1
    # via
    #   -r requirements.txt
    #   tablib
xlwt==1.3.0
    # via
    #   -r requirements.txt
    #   tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
